Create and format search bar
Wrote the HTML for the search bar, google logo, search/feeling
lucky buttons and the talk to text button with microphone logo.
Formatted these items in CSS to mirror the actual google homepage as
closely as possible. Was unable to complete the magnifiying glass
image on the left side of the search bar as I am unfamiliar with svg
tags in HTML5.